Create only a single query and mutation class
